+++ app/controllers/oembed_controller.rb
@@ -16,13 +16,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+require 'cgi'
+
 class OembedController < ApplicationController
   respond_to :json, :xml
 
   def show
     not_found if params[:url].blank?
 
-    url = Rails.application.routes.recognize_path(params[:url])
+    url = CGI.unescape(params[:url])
+    url = Rails.application.routes.recognize_path(url)
+
     id_hash = Article.from_uri(url[:id])
     article = ArticleDecorator.where(id_hash).includes(:retrieval_statuses).first